put error handling and error definitions in one file named errors.py
